fix: update openzeppelin version
Since openzeppelin-contracts-upgradeable v5.0.0, the DoubleEndedQueue.sol contract has been removed from contract/utils/structs/DoubleEndedQueue.sol. This means that when trying to use the Greenfield Contract library, an error will be displayed. 

However, the DoubleEndedQueue.sol contract is still available in openzeppelin-contracts v4.9.3 and v5.0.0, 

as well as in openzeppelin-contracts-upgradeable v4.9.3. However, it is not available inopenzeppelin-contracts-upgradeable v5.0.0.